Refactoring Moved FeatureVector to Double instead of float Basic implementation of training of features instead of double arrays

+++ src/main/java/de/thatsich/openfx/featureextraction/api/control/entity/IFeatureVector.java
@@ -9,7 +9,7 @@ import javafx.beans.property.ReadOnlyListProperty;
  */
 public interface IFeatureVector
 {
-	ReadOnlyListProperty<Float> vector();
+	ReadOnlyListProperty<Double> vector();
 
 	ReadOnlyBooleanProperty isPositive();
 }
